Ensure query dates are availables for composite requests
