+++ frontend/src/components/MapView/DateSelector/index.test.tsx
@@ -3,7 +3,6 @@ import { render } from '@testing-library/react';
 import configureStore from 'redux-mock-store';
 import { PanelSize } from 'config/types';
 import DateSelector from '.';
-import { findClosestDate } from './utils';
 
 jest.mock('react-router-dom', () => ({
   useHistory: () => ({
@@ -54,25 +53,3 @@ test('renders as expected', () => {
   // eslint-disable-next-line fp/no-mutation
   global.Date.now = realDateNow;
 });
-
-test('DateSelector utils', () => {
-  const findClosestDateResult = new Date(
-    findClosestDate(
-      findClosestDateData.date,
-      findClosestDateData.availableDates,
-    ),
-  ).toISOString();
-
-  expect(findClosestDateResult).toBe(findClosestDateData.result);
-});
-
-const findClosestDateData = {
-  date: 1702288800000,
-  availableDates: [
-    1689076800000, 1689940800000, 1690891200000, 1691755200000, 1692619200000,
-    1693569600000, 1694433600000, 1695297600000, 1696161600000, 1697025600000,
-    1697889600000, 1698840000000, 1699704000000, 1700568000000, 1701432000000,
-    1702296000000, 1703160000000, 1704110400000, 1704974400000,
-  ],
-  result: '2023-12-11T12:00:00.000Z',
-};
