+++ frontend/src/context/layers/composite_data.ts
@@ -4,7 +4,11 @@ import type { CompositeLayerProps } from 'config/types';
 import { fetchWithTimeout } from 'utils/fetch-with-timeout';
 import { HTTPError, LocalError } from 'utils/error-utils';
 import { addNotification } from 'context/notificationStateSlice';
-import { getFormattedDate, getSeasonBounds } from 'utils/date-utils';
+import {
+  findClosestDate,
+  getFormattedDate,
+  getSeasonBounds,
+} from 'utils/date-utils';
 
 import type { LayerDataParams, LazyLoader } from './layer-data';
 
@@ -13,9 +17,11 @@ export interface CompositeLayerData extends FeatureCollection {}
 export const fetchCompositeLayerData: LazyLoader<CompositeLayerProps> =
   () =>
   async (params: LayerDataParams<CompositeLayerProps>, { dispatch }) => {
-    const { layer, date } = params;
+    const { layer, date, availableDates } = params;
+
     const referenceDate = date ? new Date(date) : new Date();
-    const seasonBounds = getSeasonBounds(referenceDate);
+    const providedSeasons = layer.validity?.seasons;
+    const seasonBounds = getSeasonBounds(referenceDate, providedSeasons);
     if (!seasonBounds) {
       console.error(
         `No season bounds found for ${layer.id} with date ${referenceDate}`,
@@ -28,7 +34,18 @@ export const fetchCompositeLayerData: LazyLoader<CompositeLayerProps> =
     // For seasonal, setting an end date to the end of the season
     const endDate = useMonthly
       ? new Date(startDate).setMonth(startDate.getMonth() + 1)
-      : seasonBounds.end;
+      : new Date(seasonBounds.end).getTime();
+
+    const availableQueryDates = availableDates
+      ? Array.from(new Set(availableDates.map(dateItem => dateItem.queryDate)))
+      : [];
+
+    const closestDateToStart = availableDates
+      ? findClosestDate(startDate.getTime(), availableQueryDates)
+      : startDate;
+    const closestDateToEnd = availableDates
+      ? findClosestDate(endDate, availableQueryDates)
+      : endDate;
 
     const {
       baseUrl,
@@ -40,8 +57,8 @@ export const fetchCompositeLayerData: LazyLoader<CompositeLayerProps> =
 
     // docs: https://hip-service.ovio.org/docs#/default/run_q_multi_geojson_q_multi_geojson_post
     const body = {
-      begin: getFormattedDate(startDate, 'default'),
-      end: getFormattedDate(endDate, 'default'),
+      begin: getFormattedDate(closestDateToStart, 'default', true),
+      end: getFormattedDate(closestDateToEnd, 'default', true),
       area: {
         min_lon: boundingBox[0],
         min_lat: boundingBox[1],
